refactor: split subagent types and formatting
